Put SelfieTestExecutionListener into its own file.
